image-gltf-processorモジュールの追加と設定の更新
- 新しいモジュール `image-gltf-processor.ts` を追加しました。このモジュールは、画像のリサイズ、メタデータ取得、GLTFまたはGLBファイルへの変換、および変換されたファイルの保存を行います。
- `main.ts` を修正し、`image-gltf-processor` モジュールの関数をインポートして使用するようにしました。
- `tsconfig.json` を更新し、型定義ファイルの生成を有効にしました。
- `package.json` を更新し、`build` スクリプトに `image-gltf-processor` モジュールのミニファイ処理を追加しました。
